requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-mantine-components==0.12.1
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.2
google-ai-generativelanguage==0.4.0
google-api-core==2.16.1
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
importlib_metadata==7.0.1
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
markupsafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==4.1.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.0
pure_eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic-core==2.14.6
pygments==2.17.2
requests==2.31.0
retrying==1.3.4
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
tenacity==8.2.3
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
wcwidth==0.2.13
werkzeug==3.0.1
wheel==0.41.2